examples/requirements.txt
python-dotenv
asyncio==3.4.3
aiofiles==23.2.1
aiohttp==3.9.5